add support network

+++ src/GameBoard/History.java
@@ -9,45 +9,46 @@ import java.util.List;
  * Date: 05.08.13
  */
 public class History {
-    private class Step{
+    private class Step {
         public int x;
         public int y;
         public int mark;
-        public Step(int x,int y,int mark){
-            this.x=x;
-            this.y=y;
-            this.mark=mark;
+
+        public Step(int x, int y, int mark) {
+            this.x = x;
+            this.y = y;
+            this.mark = mark;
         }
     }
 
-    private List<Step> steps=new ArrayList<Step>();
+    private List<Step> steps = new ArrayList<Step>();
 
-    public void append(int x,int y,int mark){
-        steps.add(new Step(x,y,mark));
+    public void append(int x, int y, int mark) {
+        steps.add(new Step(x, y, mark));
     }
 
-    public void goBackToStep(int numStep,int[][] board){
-        if(numStep<steps.size() && numStep>=0){
-            for(int i=numStep+2;i<steps.size();i++){
-                board[steps.get(i).y][steps.get(i).x]=GameBoard.EMPTY;
+    public void goBackToStep(int numStep, byte[][] board) {
+        if (numStep < steps.size() && numStep >= 0) {
+            for (int i = numStep + 2; i < steps.size(); i++) {
+                board[steps.get(i).y][steps.get(i).x] = GameBoard.EMPTY;
                 steps.remove(i);
             }
         }
     }
 
-    public void print(){
-        int st=1;
-        for(int i=0;i<steps.size();i++){
-            if(i % 2==0){
-                System.out.printf("%d:\n",st);
+    public void print() {
+        int st = 1;
+        for (int i = 0; i < steps.size(); i++) {
+            if (i % 2 == 0) {
+                System.out.printf("%d:\n", st);
                 st++;
             }
             System.out.printf(
                     "\t%c: %d,%d\n",
-                    (steps.get(i).mark==GameBoard.CROSS) ? 'X' : '0',
+                    (steps.get(i).mark == GameBoard.CROSS) ? 'X' : '0',
                     steps.get(i).x,
                     steps.get(i).y
-                    );
+            );
         }
     }
 }
